fixed ui bug for total sold and cost of free download in releaseCreateConfirm
